Properly vendor upstream dependencies.
distributed-closure was removed as a vendored module, since it's been
released to Hackage now, so can just be an extra-dep.
binary-serialise-cbor was added using Stack's dedicated support for
Git repos.

This is far more robust that these repos have been checked out in some
location relative to the current repo, but outside of it.