Use inline-c to implement Foreign.Java. Expand list of covered JNI bindings.
The main reason for using inline-c is that it's ultimately the more
typesafe approach. It's the only approach, along with c2hs, that
checks for inconsistencies between the Haskell types of the bound
functions and the C types declared in the .h files. A nice side
benefit is that cbits/Java.c becomes unnecessary, since any helper
code is now inline in the .hs file, not in a separate file whose type
signatures must be kept in sync with the bindings.