Deal with serializing/deserializing functions (nearly) properly.
This amounts to a fairly hefty rewrite of C.D.S.Closure. The main
feature here is that we no longer need to peg `invoke` to functions of
a single concrete type, say `String -> Bool`. We can now uniformly
create a Java function object polymorphically. Well, nearly
polymorphically - limitations of the current implementation
of -XStaticPointers mean that to really make it happen we'd need some
healy unsafeCoerce'ing of reified dictionaries.

Not entirely impossible, but so for now we do a hacky dispatch on the
TypeRep inside the instance. Hopefully should disappear soon.