Get rid of the env parameter to all Java functions.
This extra parameter, which could have been hidden behind a reader monad
but at the cost of forcing all code, including pure code, to use
a monadic style, is now stored in a thread-local variable, using thread
local storage (TLS). We use the `thread-local-storage` package to achiev
that. Indeed, within a single thread, the env parameter is constant. So
instead of threading it through implicitly, we just don't thread it
through at all.

This patch further reduces the risk of tricky to resolve bugs that would
arise from one thread mistakenly sending the env parameter to another
thread, e.g. via a closure.