Remove the need to build apps as shared libraries.
Instead, we build them as position independent executables (PIE).
These executables look, in reality, just like any regular shared
library. But it's a bona fide executable, with a main entry point.
Since this entry point is exported as a symbol, we can jump control
flow directly to it from Java.

A side benefit of this is that we no longer have to perform GHC RTS
initialization ourselves, because we call the main() function
generated by GHC, which does whatever it needs to do to execute the
program and clean up.

Because we call the main() function, which necessarily has a standard
type signature, our entry point into the Haskell code can no longer
accept a JNIEnv as an argument. So we introduce a tiny bootstrapping
bit of glue code that takes the JNIEnv supplied by Java, grabs it's
JavaVM structure and sticks it in a public global variable. That's
quite okay, because the current JNISpec says only one JVM per address
space is supported anyways. So it's not like the presence of this
global variable precludes us from supporting multiple JVM's in the
same process - the JVM doesn't support it anyways.