Package dynamically linked executables dependencies in .jar.
We arrange for maven to discover all dynamically linked libraries that
the Haskell app depends on, and stick 'em in the jar. For simplicity,
we encapsulate all the discovered libs in a .zip file that then goes
into the JAR. This gives us an easy manifest of items to unpack at
runtime, which we do with some Java code, but we could consider other
approaches in the future.