Set origin processing flag in ld, and fix quoting of $ORIGIN.
We add $ORIGIN as the RPATH of our apps to tell the linker that if the
dynamic libraries they depend upon are to be found in the same folder,
then just use that. However, it turns out that some systems (but not
all!) simply ignore $ORIGIN unless the object is "marked as requiring
origin processing". You do so by linking with the -Wl,-z,origin option.

Now it turns out that these systems furthermore are rather pedantic
about the quoting of $ORIGIN: there should be no quotes. We said
'$ORIGIN' instead of $ORIGIN since that's the right way to avoid shell
escapes, but in a .cabal file that's one level of quoting too far...