Add `new`/`call`: higher-level wrappers around `newObject`/`callObjectMethod`.
These use the arguments types and target type to infer a class/method
signature, so they don't have to be repeated by hand. Obviates the need
to call `findClass` directly before calling a method or creating a new
instance.

This resulted in a substantial reduction in the number of lines of code
in the existing bindings.