Have Cabal build and package sparkle class files.
We use a custom build-type and hook a call to Gradle in Setup.hs. In
this way, class files get build as part of `stack build`, not just
haskell source files. Furthermore, the resulting .jar is installed as
a data-file. Since this .jar is now in a known location, we can now
reuse it in out-of-tree builds.

We use Gradle for the build, rather than Maven. Simpler.