Introduce JNI.String type. Switch API to using it instead of ByteString.
ByteStrings and JNI strings don't have the same invariants. Crucially,
ByteString values are not necessarily null terminated. We used to hack
around that in the implementation of `call` and `callStatic`, but it's
cleaner to ask the user to provide us with evidence that provided string
really is null-terminated. We can do that with a newtype around
`CString`.

The newtype is necessary because we want to enfore another invariant:
JNI strings are always UTF-8 encoded, irrespective of the locale.

To boot, we have a more efficient way of building signatures, i.e.
JNI.String's, based on ByteString.Builder.