Use bytestrings as the backing store for JNI.String.
It's now a wrapper around bytestrings, rather than manipulating
ForeignPtr's directly. This turns out to make the implementation quite
a bit shorter, and also avoids any problems with creating a new
foreign pointer to the same buffer as a bytestring. That is a recipe
for disaster: the bytestring's finalizer will eventually free the
buffer, and we'll be left high and dry pointing to unallocated memory.