Use Nix to provision a now complete Docker container.
The previous Docker container for the project included Java, but not
Spark. So while it was enough to *build* sparkle, it wasn't enough to
run Spark applications built using sparkle. This new Docker image
contains everything that is needed to both build and run apps.

In order to avoid the Nix shell.nix and the Docker Dockerfile from
drifting away from each other over time, we actually provision nearly
all system dependencies in the image using Nix. However, we *don't*
use Stack's --nix support nested inside Stack's --docker support.
Instead, we crafted a special entrypoint for the container that makes
Stack oblivious to the very existence of Nix. Stack just blithely
grabs stuff from the current PATH, however it was provided to it by
Nix.