mapPartitions and mapPartitionsWithIndex.
Uses the new `jvm-streaming` package under the hood. Note that these
functions are definable even without depending on `jvm-streaming` at
all. But only usable if the instances defined in `jvm-streaming` are in
scope. Since that package can only build on GHC >= 8.0.2, it means we
intentionally export the same API for all compiler versions (a good
thing), even if some of the API can't be used without a newer compiler.