Adapt to simpler type of Reify/Reflect.
Now that Uncurry is gone, we needed a new solution for the Closure
instances. I was intent on keeping the currying of closure types,
rather than using newtype wrappers for each arity, which ends up being
terribly verbose. To this end, we removed the existing Reify/Reflect
instances for Closure entirely, and instead defined a specialized
class that takes into account the target arity when reflecting.

reflectFun.