Preliminary bazel support.
For the most part this should be self-explanatory. There is a
work-around for `Paths_` and I'm not quite happy that it works:

* `Sparkle.hs` currently relies on `Paths_` to find sparkle jar which
  is a Cabal feature. I have put in a workaround where cabal users can
  keep relying on it but others can use a different `main` which
  requires that this is passed in.
* Trying to increase UX for the above build target, I have writted a
  small wrapper around the binary which fills in this flag
  for the user with location of the jar produced earlier on in the
  build. This way user does not manually fiddle with paths.
* This works if we `bazel build` and `./bazel-bin/sparkle` but doesn't
  work if we try to `bazel run`. I could not figure out how to get the
  latter working.

Another note is that we have to list all dependencies used rather than
getting transitive dependencies from a CLASSPATH. This is due to a bug
in bazel not allowing us to pull CLASSPATH reliably from jars.